requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-multipart==0.0.20
PyPDF2==3.0.1
sentence-transformers==3.3.1
transformers
rank_bm25
python-dotenv
faiss-cpu==1.9.0.post1
google-generativeai
numpy==1.26.4
torch==2.5.1
